requirements.txt
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
sentence-transformers==3.3.1
groq>=0.11.0
kagglehub==0.3.4
torch==2.4.1
transformers==4.45.2
python-dotenv>=0.19.0,<2.0.0
httpx>=0.27.0
streamlit>=1.28.0